+++ index.js
@@ -1,58 +1,14 @@
 #!/usr/bin/env node
-"use strict";
-const commander  = require("commander");
-const request    = require("request-promise");
-const version    = require("./package.json").version;
-const open       = require("open");
-const redditUrls = [
-        "https://www.reddit.com/r/funnyvideos/.json",
-        "https://www.reddit.com/r/Funnypics/.json",
-        "https://www.reddit.com/r/funnygifs/.json",
-        "https://www.reddit.com/r/Funny/.json"
-      ];
-  //    Register arguments with commander
-  commander.version(version)
-            .option('-v, --video', 'opens a funny video')
-            .option('-g, --gif',   'opens a funny gif')
-            .option('-p, --pic',   'opens a funny pic')
-            .option('-q, --quiet', 'opens a funny not video')
-            .option('-o, --output','output only without opening')
-            .parse(process.argv);
+var exec = require('child_process').exec;
+var path = require('path');
 
-  //  Sets the lookup url based on command line arguments
-  let funUrl = redditUrls[3];
-  if( commander.video ){
-    funUrl = redditUrls[0];
-  }
-  if( commander.pic ){
-    funUrl = redditUrls[1];
-  }
-  if( commander.gif ){
-    funUrl = redditUrls[2];
-  }
-  if( commander.quiet ){
-    funUrl = redditUrls[Math.floor(Math.random()*redditUrls.length)+1];
-  }
+var relPath = __dirname + '/make-me-lol.js';
+relPath = relPath.replace(/\\/g, '/');
 
-  console.log('Get ready to lol...');
-  //  Go get our JSON data
-  request(funUrl)
-    .then( data => {
-      let parsed   = JSON.parse(data).data.children,
-          l = parsed.length,
-          n = Math.floor(Math.random() * (l - 1)) + 1;
-      if( commander.quiet && parsed[n].data.domain.indexOf("youtube") > -1 ){
-        n = Math.floor(Math.random() * (l - 1)) + 1;
-      };
-      //  open our funny url
-      if( !commander.output ){
-        open(parsed[n].data.url);
-      }else{
-        console.log("Your lol can be found over here:\n"+parsed[n].data.url);
-      }
-      process.exit(1);
-    })
-    .catch( err => {
-      console.warn(err);
-      process.exit(0);
-    });
+exec('node --harmony ' + relPath, function(error, stdout, stderr) {
+    console.log(stdout);
+    console.error(stderr);
+    if (error) {
+        console.error(error);
+    }
+});
